\name{mergeData}
\docType{instance}
\alias{mergeData}
\alias{sample1}
\alias{sample2}
\alias{sample3}
\title{MergeMaid instance data for merging}
\description{
  These are 3 artificial data sets, generated using rnorm(), for the purpose
  of illustrating the package.  All are list objects.  For some of the
  "genes",  the expression values in the 3 datasets are not independent.
}
\usage{data(mergeData)}

\keyword{methods}
